backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
djangorestframework-simplejwt>=5.3,<5.4
psycopg[binary]>=3.2,<3.3
